feat: Redis实现非原子性去重

+++ snorlax-handler/src/main/java/com/wyu/snorlax/deduplication/builder/DeduplicationParam.java
@@ -19,10 +19,10 @@ public class DeduplicationParam {
     /**
      * 去重时间窗
      */
-    private Long time;
+    private int time;
 
     /**
      * 窗内去重次数阈值
      */
-    private Integer count;
+    private int count;
 }
